+++ features/fixtures/formatters/passed-rule.json.ts
@@ -1,0 +1,37 @@
+module.exports = [
+  {
+    description: '',
+    elements: [
+      {
+        description: '',
+        id: 'a-feature;a-rule;an-example',
+        keyword: 'Example',
+        line: 3,
+        name: 'an example',
+        steps: [
+          {
+            arguments: [],
+            keyword: 'Given ',
+            line: 4,
+            match: {
+              location: 'features/step_definitions/steps.js:3',
+            },
+            name: 'a step',
+            result: {
+              duration: 0,
+              status: 'passed',
+            },
+          },
+        ],
+        tags: [],
+        type: 'scenario',
+      },
+    ],
+    id: 'a-feature',
+    keyword: 'Feature',
+    line: 1,
+    name: 'a feature',
+    tags: [],
+    uri: 'features/a.feature',
+  },
+]