+++ CHANGELOG.md
@@ -2,6 +2,10 @@ Please see [CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CO
 
 ### [Unreleased](https://github.com/cucumber/cucumber-js/compare/v6.0.5...master) (In Git)
 
+#### New Features
+
+* Add support for Gherkin's [Rule/Example syntax](https://cucumber.io/docs/gherkin/reference/#rule)
+
 #### Breaking changes
 
 * Drop support for Node.js 8